Summary of Release 0.4
Fixed bug where the default option for TxRequests was incorrectly set to the disble ACK option.  Because of this bug you were not getting ACK errors.  Thanks to Terrence at UVA for reporting this.

Improved error handling:  All software errors related to parsing response packets are now returned as an ErrorResponse.  
As a result of this change I removed the getErrorList method in the XBee class

waitForResponse(long timeout) now throws a XBeeTimeoutException if timeout is reached

Added new and clarified existing JavaDoc documentation

Renamed the setTimeout method in the XBee class to setSendSynchronousTimeout, to indicate it only applies to the synchronousSend method.

git-svn-id: http://xbee-api.googlecode.com/svn/trunk@65 e138a6d8-124a-0410-b381-050080eecf1f